requirements.txt
python-telegram-bot>=20.0
openai>=1.0.0
anthropic>=0.18.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
yt-dlp>=2023.12.30
pypdf>=3.17.0
python-dotenv>=1.0.0
google-cloud-vision>=3.4.0
gnews>=0.3.0
tavily-python>=0.3.0
python-dateutil>=2.8.0
aiohttp>=3.9.0
youtube-transcript-api>=0.6.1
Pillow>=10.0.0
